Вынес рендер и шаблоны тоже в отдельные файлы.
